# grpct/proxy/docs/swagger.yaml
basePath: /
definitions:
  models.Address:
    properties:
      area: {}
      area_fias_id: {}
      area_kladr_id: {}
      area_type: {}
      area_type_full: {}
      area_with_type: {}
      beltway_distance: {}
      beltway_hit: {}
      block: {}
      block_type: {}
      block_type_full: {}
      capital_marker:
        type: string
      city:
        type: string
      city_area: {}
      city_district: {}
      city_district_fias_id: {}
      city_district_kladr_id: {}
      city_district_type: {}
      city_district_type_full: {}
      city_district_with_type: {}
      city_fias_id:
        type: string
      city_kladr_id:
        type: string
      city_type:
        type: string
      city_type_full:
        type: string
      city_with_type:
        type: string
      country:
        type: string
      country_iso_code:
        type: string
      divisions: {}
      entrance: {}
      federal_district: {}
      fias_actuality_state:
        type: string
      fias_cadastre_number:
        type: string
      fias_id:
        type: string
      fias_level:
        type: string
      flat: {}
      flat_area: {}
      flat_cadnum: {}
      flat_fias_id: {}
      flat_price: {}
      flat_type: {}
      flat_type_full: {}
      floor: {}
      geo_lat:
        type: string
      geo_lon:
        type: string
      geoname_id:
        type: string
      history_values:
        items:
          type: string
        type: array
      house: {}
      house_cadnum: {}
      house_fias_id: {}
      house_kladr_id: {}
      house_type: {}
      house_type_full: {}
      kladr_id:
        type: string
      metro: {}
      okato:
        type: string
      oktmo:
        type: string
      postal_box: {}
      postal_code: {}
      qc: {}
      qc_complete: {}
      qc_geo:
        type: string
      qc_house: {}
      region:
        type: string
      region_fias_id:
        type: string
      region_iso_code:
        type: string
      region_kladr_id:
        type: string
      region_type:
        type: string
      region_type_full:
        type: string
      region_with_type:
        type: string
      source: {}
      square_meter_price: {}
      stead: {}
      stead_cadnum: {}
      stead_fias_id: {}
      stead_type: {}
      stead_type_full: {}
      street:
        type: string
      street_fias_id:
        type: string
      street_kladr_id:
        type: string
      street_type:
        type: string
      street_type_full:
        type: string
      street_with_type:
        type: string
      tax_office:
        type: string
      tax_office_legal:
        type: string
      timezone: {}
      unparsed_parts: {}
    type: object
  models.SearchRequest:
    properties:
      query:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the API documentation for the Proxy Service.
  title: Proxy Service API
  version: "1.0"
paths:
  /api/address/geocode:
    post:
      consumes:
      - application/json
      description: Handles geocode requests
      operationId: geocode-handler
      parameters:
      - description: Latitude
        in: body
        name: lat
        required: true
        schema:
          type: number
      - description: Longitude
        in: body
        name: lng
        required: true
        schema:
          type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Address'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Geocode handler
  /api/address/search:
    post:
      consumes:
      - application/json
      description: Handles search requests
      operationId: search-handler
      parameters:
      - description: Search query
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/models.SearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Address'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Search handler
swagger: "2.0"